refactor(inventory): decompose god-files + extract @pops/ui composites
Closes #2038 — Decompose LocationTreePage (1007→129 lines), ItemFormPage
(909→149 lines), and ItemDetailPage (750→131 lines) into page-shell +
sections + hooks pattern. All tRPC calls moved to useXPageModel hooks.
Remove 3 per-file oxlint max-lines overrides now that limits are met.

Closes #2048 — Extract ContainerPanel to @pops/ui
Closes #2026 — Use ContainerPanel in LocationContentsPanel

Closes #2047 — Extract RelatedItemsList to @pops/ui
Closes #2027 — Use RelatedItemsList in ConnectionsList

Closes #2045 — Extract SearchPickerDialog to @pops/ui
Closes #2018 — Use SearchPickerDialog in ConnectDialog + LinkDocumentDialog

Closes #1996 — Replace ad-hoc stat tiles in DashboardWidgets with StatCard
(adds onClick support to StatCard for the Warranties tile navigation)
