+++ packages/ui/src/index.ts
@@ -94,6 +94,11 @@ export * from './components/Select';
 export * from './components/StatCard';
 export * from './components/TextInput';
 
+// Shared composites
+export * from './components/ContainerPanel';
+export * from './components/RelatedItemsList';
+export * from './components/SearchPickerDialog';
+
 // Layout composites
 export * from './components/PageHeader';
 export * from './components/ViewToggleGroup';
